+++ src/main/java/appeng/worldgen/meteorite/settings/CraterLakeState.java
@@ -1,4 +1,4 @@
-package appeng.worldgen.meteorite;
+package appeng.worldgen.meteorite.settings;
 
 public enum CraterLakeState {
     UNSET,
